% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Example data}
\format{
\subsection{\code{sample_data}}{

A data frame with 10 observations on 2 variables.
\describe{
\item{x}{numeric vector}
\item{y}{numeric vector}
}
}
}
\source{
Manually created.
}
\usage{
sample_data
}
\description{
Example data created manually for the purpose of this package.
}
\keyword{datasets}
